refactor project, get rid of unused methods, aggregates etc
