--- src/main/kotlin/ru/quipy/delivery/api/DeliveryAggregate.kt
@@ -1,7 +1,0 @@
-package ru.quipy.delivery.api
-
-import ru.quipy.core.annotations.AggregateType
-import ru.quipy.domain.Aggregate
-
-@AggregateType(aggregateEventsTableName = "aggregate-delivery")
-class DeliveryAggregate : Aggregate
