+++ src/main/kotlin/ru/quipy/payments/logic/PaymentService.kt
@@ -43,6 +43,8 @@ data class ExternalServiceProperties(
  * Describes response from external service.
  */
 class ExternalSysResponse(
+    val transactionId: String,
+    val paymentId: String,
     val result: Boolean,
     val message: String? = null,
 )
