+++ src/main/kotlin/ru/quipy/payments/subscribers/PaymentTransactionsSubscriber.kt
@@ -26,7 +26,11 @@ class PaymentTransactionsSubscriber {
 
     @PostConstruct
     fun init() {
-        subscriptionsManager.createSubscriber(PaymentAggregate::class, "payments:payment-processings-subscriber", retryConf = RetryConf(1, RetryFailedStrategy.SKIP_EVENT)) {
+        subscriptionsManager.createSubscriber(
+            PaymentAggregate::class,
+            "payments:payment-processings-subscriber",
+            retryConf = RetryConf(1, RetryFailedStrategy.SKIP_EVENT)
+        ) {
             `when`(PaymentProcessedEvent::class) { event ->
                 paymentLog.computeIfAbsent(event.orderId) {
                     CopyOnWriteArrayList()
