+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
 server.address=0.0.0.0
 server.port=8081
 server.http2.enabled=true
-spring.mvc.async.request-timeout=300000
 spring.main.allow-bean-definition-overriding=true
 
 # MongoDB properties
@@ -12,25 +11,8 @@ spring.data.mongodb.database=online-shop
 # Tiny event sourcing library properties
 event.sourcing.auto-scan-enabled=true
 event.sourcing.scan-package=ru.quipy
-event.sourcing.snapshot-frequency=10
-
-# event stream reader props
-event.sourcing.stream-read-period=500
-event.sourcing.reader-commit-period-messages=500
-event.sourcing.reader-commit-period-millis=1000
-event.sourcing.stream-batch-size=500
-
-# events writer props
-event.sourcing.batch-enabled=true
-event.sourcing.batch-size=100
-event.sourcing.batch-period-millis=50
-#JDBC_BATCH, STORED_PROCEDURE
-event.sourcing.batch-mode=STORED_PROCEDURE
-event.sourcing.db-schema=event_sourcing_store
-
 event.sourcing.snapshots-enabled=false
 event.sourcing.sagas-enabled=false
-
 
 # Postgres event store properties
 spring.datasource.hikari.jdbc-url=jdbc:postgresql://localhost:5432/postgres
