+++ src/main/resources/logback-spring.xml
@@ -4,9 +4,6 @@
     <include resource="org/springframework/boot/logging/logback/console-appender.xml"/>
 
     <springProfile name="!dev">
-        <springProperty name="service" source="spring.application.name"/>
-        <springProperty name="logstashUrl" source="logstash.url"/>
-
         <logger name="ru.quipy" level="WARN"/>
 
         <root level="WARN">
